requirements.txt
google-api-python-client==2.190.0
pandas==3.0.1
python-dotenv==1.2.1
requests==2.32.5